Added markup of landing and started editing products
