+++ api.ts
@@ -6,8 +6,8 @@ export interface User {
   email: string;
 }
 
-export const BASE_URL = "https://tbc-r.vercel.app";
-// export const BASE_URL = "http://localhost:3000";
+// export const BASE_URL = "https://tbc-r.vercel.app";
+export const BASE_URL = "http://localhost:3000";
 // export const BASE_URL =
 //   process.env.NODE_ENV === "development"
 //     ? "http://localhost:3000"
