+++ src/app/components/BlogList.tsx
@@ -61,15 +61,6 @@ export default function BlogList({ blogs }: BlogListProps) {
             className="w-full md:w-96 px-4 py-2 rounded-md border border-gray-300 focus:outline-none focus:border-blue-500 text-black"
           />
         </div>
-        {isAdmin && (
-          <div>
-            <Link href="/admin/add-blog">
-              <button className="p-3 bg-primaryColor rounded text-white">
-                Add Blog
-              </button>
-            </Link>
-          </div>
-        )}
         <div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 gap-x-8 gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
           {filteredBlogs.map((blog) => (
             <div key={`blog-generate-${blog.id}`}>
