+++ src/app/components/CTAsection.tsx
@@ -1,6 +1,8 @@
+/* eslint-disable react/no-unescaped-entities */
 import { XMarkIcon } from "@heroicons/react/24/outline";
 import React from "react";
 import { Video } from "./Video";
+import Link from "next/link";
 // import Video from "./Video";
 
 export default function CTA() {
@@ -14,22 +16,20 @@ export default function CTA() {
               Discover Unforgettable Events Near You! <br />
             </h2>
             <p className="mx-auto mt-6 max-w-xl text-lg leading-8 text-indigo-200">
-              Incididunt sint fugiat pariatur cupidatat consectetur sit cillum
-              anim id veniam aliqua proident excepteur commodo do ea.
+              Welcome to EVENT, your go-to platform for finding and attending
+              exciting events. Whether you're into music, arts, sports, or
+              networking, we've got something for everyone!
             </p>
             <div className="mt-10 flex items-center justify-center gap-x-6">
-              <a
-                href="#"
-                className="rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold text-indigo-600 shadow-sm hover:bg-indigo-50 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white"
+              <Link
+                href="/products"
+                className="rounded-md bg-white px-3.5 py-2.5 text-sm font-semibold text-primaryColor shadow-sm hover:bg-indigo-50 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white"
               >
-                Get started
-              </a>
-              <a
-                href="#"
-                className="text-sm font-semibold leading-6 text-white"
-              >
-                Learn more <span aria-hidden="true">→</span>
-              </a>
+                View Events
+                <span className="ml-2" aria-hidden="true">
+                  →
+                </span>
+              </Link>
             </div>
           </div>
         </div>
