+++ src/app/components/Header.tsx
@@ -27,7 +27,6 @@ export default function Header() {
     { name: t("Blogs"), href: "/blogs" },
     { name: t("Contact"), href: "/contact" },
     ...(user ? [{ name: t("Profile"), href: "/profile" }] : []),
-    ...(isAdmin ? [{ name: t("Admin"), href: "/admin/add" }] : []),
   ];
 
   const [showHeader, setShowHeader] = useState(true);
@@ -68,7 +67,7 @@ export default function Header() {
     >
       {({ open }) => (
         <>
-          <div className="container mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="container mx-auto px-4 sm:px-6 lg:px-8 ">
             <div className="flex h-16 justify-between">
               <div className="flex">
                 <div className="flex flex-shrink-0 items-center">
@@ -190,6 +189,22 @@ export default function Header() {
                               )}
                             >
                               Add Product
+                            </a>
+                          )}
+                        </Menu.Item>
+                      )}
+
+                      {isAdmin && (
+                        <Menu.Item>
+                          {({ active }) => (
+                            <a
+                              href="/admin/add-blog"
+                              className={classNames(
+                                active ? "bg-gray-100" : "",
+                                "block px-4 py-2 text-sm text-gray-700"
+                              )}
+                            >
+                              Add Blog
                             </a>
                           )}
                         </Menu.Item>
