+++ src/app/components/HomeLayout.tsx
@@ -1,0 +1,312 @@
+/* eslint-disable react/no-unescaped-entities */
+"use client";
+import { useState } from "react";
+import {
+  AcademicCapIcon,
+  CheckCircleIcon,
+  HandRaisedIcon,
+  RocketLaunchIcon,
+  SparklesIcon,
+  SunIcon,
+  UserGroupIcon,
+} from "@heroicons/react/20/solid";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
+import Image from "next/image";
+import PartnerCompanies from "../components/PartnerCompanies";
+import Carousel from "../components/Carousel";
+import EventCarousel from "../components/EventCarousel";
+import CTA from "../components/CTAsection";
+import BlogCarousel from "./BlogCarousel";
+
+const navigation = [
+  { name: "Product", href: "#" },
+  { name: "Features", href: "#" },
+  { name: "Marketplace", href: "#" },
+  { name: "Company", href: "#" },
+];
+const stats = [
+  { label: "Site was founded", value: "2012" },
+  { label: "People on the team", value: "120+" },
+  { label: "Users on the platform", value: "250k" },
+  { label: "Paid out to creators", value: "$2M" },
+];
+const values = [
+  {
+    name: "Flexible Find",
+    description:
+      "Search events by date, category, or location to find exactly what interests you.",
+    icon: RocketLaunchIcon,
+  },
+  {
+    name: "RSVP securely",
+    description:
+      "Securely RSVP to events with just a few clicks, or purchase tickets hassle-free.",
+    icon: HandRaisedIcon,
+  },
+  {
+    name: "Details",
+    description:
+      "Get comprehensive event information at a glance, including ticket prices, venue details, and more.",
+    icon: UserGroupIcon,
+  },
+  {
+    name: "New Connections",
+    description:
+      "Connect with other attendees, share experiences, and build a community around your favorite events.",
+    icon: AcademicCapIcon,
+  },
+  {
+    name: "Mobile-friendly Features.",
+    description:
+      "Access our platform on the go with our mobile app for iOS and Android.",
+    icon: SparklesIcon,
+  },
+  {
+    name: "Stay Informed",
+    description:
+      "Receive personalized event recommendations and updates so you're always in the know.",
+    icon: SunIcon,
+  },
+];
+
+const benefits = [
+  "Competitive salaries",
+  "Flexible work hours",
+  "30 days of paid vacation",
+  "Annual team retreats",
+  "Benefits for you and your family",
+  "A great work environment",
+];
+const footerNavigation = {
+  solutions: [
+    { name: "Marketing", href: "#" },
+    { name: "Analytics", href: "#" },
+    { name: "Commerce", href: "#" },
+    { name: "Insights", href: "#" },
+  ],
+  support: [
+    { name: "Pricing", href: "#" },
+    { name: "Documentation", href: "#" },
+    { name: "Guides", href: "#" },
+    { name: "API Status", href: "#" },
+  ],
+  company: [
+    { name: "About", href: "#" },
+    { name: "Blog", href: "#" },
+    { name: "Jobs", href: "#" },
+    { name: "Press", href: "#" },
+    { name: "Partners", href: "#" },
+  ],
+  legal: [
+    { name: "Claim", href: "#" },
+    { name: "Privacy", href: "#" },
+    { name: "Terms", href: "#" },
+  ],
+  social: [
+    {
+      name: "Facebook",
+      href: "#",
+      icon: () => (
+        <svg fill="currentColor" viewBox="0 0 24 24">
+          <path
+            fillRule="evenodd"
+            d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: "Instagram",
+      href: "#",
+      icon: () => (
+        <svg fill="currentColor" viewBox="0 0 24 24">
+          <path
+            fillRule="evenodd"
+            d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: "X",
+      href: "#",
+      icon: () => (
+        <svg fill="currentColor" viewBox="0 0 24 24">
+          <path d="M13.6823 10.6218L20.2391 3H18.6854L12.9921 9.61788L8.44486 3H3.2002L10.0765 13.0074L3.2002 21H4.75404L10.7663 14.0113L15.5685 21H20.8131L13.6819 10.6218H13.6823ZM11.5541 13.0956L10.8574 12.0991L5.31391 4.16971H7.70053L12.1742 10.5689L12.8709 11.5655L18.6861 19.8835H16.2995L11.5541 13.096V13.0956Z" />
+        </svg>
+      ),
+    },
+    {
+      name: "GitHub",
+      href: "#",
+      icon: () => (
+        <svg fill="currentColor" viewBox="0 0 24 24">
+          <path
+            fillRule="evenodd"
+            d="M12 2C6.477 2 2 6.484 2 12.017c0 4.425 2.865 8.18 6.839 9.504.5.092.682-.217.682-.483 0-.237-.008-.868-.013-1.703-2.782.605-3.369-1.343-3.369-1.343-.454-1.158-1.11-1.466-1.11-1.466-.908-.62.069-.608.069-.608 1.003.07 1.531 1.032 1.531 1.032.892 1.53 2.341 1.088 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.113-4.555-4.951 0-1.093.39-1.988 1.029-2.688-.103-.253-.446-1.272.098-2.65 0 0 .84-.27 2.75 1.026A9.564 9.564 0 0112 6.844c.85.004 1.705.115 2.504.337 1.909-1.296 2.747-1.027 2.747-1.027.546 1.379.202 2.398.1 2.651.64.7 1.028 1.595 1.028 2.688 0 3.848-2.339 4.695-4.566 4.943.359.309.678.92.678 1.855 0 1.338-.012 2.419-.012 2.747 0 .268.18.58.688.482A10.019 10.019 0 0022 12.017C22 6.484 17.522 2 12 2z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+    {
+      name: "YouTube",
+      href: "#",
+      icon: () => (
+        <svg fill="currentColor" viewBox="0 0 24 24">
+          <path
+            fillRule="evenodd"
+            d="M19.812 5.418c.861.23 1.538.907 1.768 1.768C21.998 8.746 22 12 22 12s0 3.255-.418 4.814a2.504 2.504 0 0 1-1.768 1.768c-1.56.419-7.814.419-7.814.419s-6.255 0-7.814-.419a2.505 2.505 0 0 1-1.768-1.768C2 15.255 2 12 2 12s0-3.255.417-4.814a2.507 2.507 0 0 1 1.768-1.768C5.744 5 11.998 5 11.998 5s6.255 0 7.814.418ZM15.194 12 10 15V9l5.194 3Z"
+            clipRule="evenodd"
+          />
+        </svg>
+      ),
+    },
+  ],
+};
+const blogPosts = [
+  {
+    id: 1,
+    title: "Vel expedita assumenda placeat aut nisi optio voluptates quas",
+    href: "#",
+    description:
+      "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
+    imageUrl:
+      "https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80",
+    date: "Mar 16, 2020",
+    datetime: "2020-03-16",
+    author: {
+      name: "Michael Foster",
+      imageUrl:
+        "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    },
+  },
+  {
+    id: 1,
+    title: "Vel expedita assumenda placeat aut nisi optio voluptates quas",
+    href: "#",
+    description:
+      "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
+    imageUrl:
+      "https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80",
+    date: "Mar 16, 2020",
+    datetime: "2020-03-16",
+    author: {
+      name: "Michael Foster",
+      imageUrl:
+        "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    },
+  },
+
+  {
+    id: 1,
+    title: "Vel expedita assumenda placeat aut nisi optio voluptates quas",
+    href: "#",
+    description:
+      "Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.",
+    imageUrl:
+      "https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80",
+    date: "Mar 16, 2020",
+    datetime: "2020-03-16",
+    author: {
+      name: "Michael Foster",
+      imageUrl:
+        "https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80",
+    },
+  },
+  // More posts...
+];
+
+export default function HomeLayout({ products, blogs }: any) {
+  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
+  return (
+    <div className="">
+      <Header />
+      <CTA />
+      <main className="">
+        {/* Background */}
+        <div
+          className="absolute inset-x-0 top-4 -z-10 flex transform-gpu justify-center overflow-hidden blur-3xl"
+          aria-hidden="true"
+        >
+          <div
+            className="aspect-[1108/632] w-[69.25rem] flex-none bg-gradient-to-r from-[#80caff] to-[#4f46e5] opacity-25"
+            style={{
+              clipPath:
+                "polygon(73.6% 51.7%, 91.7% 11.8%, 100% 46.4%, 97.4% 82.2%, 92.5% 84.9%, 75.7% 64%, 55.3% 47.5%, 46.5% 49.4%, 45% 62.9%, 50.3% 87.2%, 21.3% 64.1%, 0.1% 100%, 5.4% 51.1%, 21.4% 63.9%, 58.9% 0.2%, 73.6% 51.7%)",
+            }}
+          />
+        </div>
+
+        {/* <div className="container mx-auto px-4  sm:px-6 lg:px-8 font-bold text-3xl text-gray-600 flex justify-center mt-10">
+          <h1>"Experience the Extraordinary Where Moments Become Memories"</h1>
+        </div> */}
+        <div className="text-center py-8 bg-gray-100 border-t-2 border-b-2 border-gray-300">
+          <h2 className="text-4xl font-bold uppercase tracking-wide text-gray-800 mb-2">
+            Experience the Extraordinary
+          </h2>
+          <h3 className="text-2xl  text-gray-600">
+            Where Moments Become Memories
+          </h3>
+        </div>
+        {/* Header section */}
+        <div className="flex flex-col sm:flex sm:flex-col md:flex md:flex-row  mt-16 container mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="mx-auto max-w-2xl pt-4 text-center">
+            <h2 className="text-4xl font-bold tracking-tight text-black sm:text-6xl font-sans">
+              We love creators
+            </h2>
+            <p className="mt-6 text-lg leading-8 text-gray-300 font-sans">
+              Anim aute id magna aliqua ad ad non deserunt sunt. Qui irure qui
+              lorem cupidatat commodo. Elit sunt amet fugiat veniam occaecat
+              fugiat aliqua.
+            </p>
+          </div>
+          <Carousel />
+        </div>
+        <div className="p-10 container mx-auto px-4   sm:px-6 lg:px-8 ">
+          <h1 className="font-bold font-sans text-lg">Latest Events</h1>
+          <EventCarousel products={products} />
+        </div>
+
+        {/* Blog section */}
+        <div className="p-10 container mx-auto px-4   sm:px-6 lg:px-8 ">
+          <h1 className="font-bold font-sans text-lg">Latest Blogs</h1>
+          <BlogCarousel blogs={blogs} />
+        </div>
+
+        {/* Values section */}
+        <div className="mt-6 sm:mt-6 container mx-auto px-4 sm:px-6 lg:px-8">
+          <div className="mx-auto  lg:mx-0">
+            <h2 className="text-3xl font-bold tracking-tight text-black sm:text-4xl">
+              Our Values
+            </h2>
+          </div>
+          <dl className="mx-auto grid  grid-cols-1 gap-8 text-base leading-7 text-gray-600 sm:grid-cols-2  lg:gap-x-16">
+            {values.map((value) => (
+              <div
+                key={`value-generate-${value.name}`}
+                className="relative pl-9"
+              >
+                <dt className="inline font-semibold text-black">
+                  <value.icon
+                    className="absolute left-1 top-1 h-5 w-5 text-primaryColor"
+                    aria-hidden="true"
+                  />
+                  {value.name}
+                </dt>{" "}
+                <dd className="inline">{value.description}</dd>
+              </div>
+            ))}
+          </dl>
+        </div>
+
+        <PartnerCompanies />
+      </main>
+
+      <Footer />
+    </div>
+  );
+}
